+++ packages/terra-date-picker/src/terra-dev-site/doc/example/DatePickerFilterDates.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Field from 'terra-form-field';
-// eslint-disable-next-line import/no-unresolved, import/extensions
+import moment from 'moment';
+// eslint-disable-next-line import/no-unresolved, import/extensions, import/no-extraneous-dependencies
 import DatePicker from 'terra-date-picker/lib/DatePicker';
 
 const propTypes = {
@@ -50,8 +51,14 @@ DatePickerExample.propTypes = propTypes;
 DatePickerExample.defualtProps = defaultProps;
 
 const isWeekday = (date) => {
-  const day = date.day();
-  return day !== 0 && day !== 6;
+  const momentDate = moment(date);
+
+  if (momentDate && momentDate.isValid()) {
+    const day = momentDate.day();
+    return day !== 0 && day !== 6;
+  }
+
+  return true;
 };
 
 const DatePickerExampleFilterDates = () => (