+++ packages/terra-date-time-picker/src/terra-dev-site/doc/example/DateTimePickerFilterDates.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import moment from 'moment';
 import PropTypes from 'prop-types';
 import Field from 'terra-form-field';
 // eslint-disable-next-line import/no-unresolved, import/extensions
@@ -50,8 +51,14 @@ DateTimePickerExample.propTypes = propTypes;
 DateTimePickerExample.defualtProps = defaultProps;
 
 const isWeekday = (date) => {
-  const day = date.day();
-  return day !== 0 && day !== 6;
+  const momentDate = moment(date);
+
+  if (momentDate && momentDate.isValid()) {
+    const day = momentDate.day();
+    return day !== 0 && day !== 6;
+  }
+
+  return true;
 };
 
 const DateTimePickerExampleFilterDates = () => (