+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -4,6 +4,17 @@ Changelog
 Unreleased
 ----------
 
+3.0.0 - (May 13, 2019)
+------------------
+### Breaking Changes
+* Removed focus-trap from abstract modal
+* Replaced with functionality that will disable focus on elements outside of the modal (not including content rendered from the modal in a portal)
+* Removed `isFocused` prop
+* Removed `fallbackFocus` prop. Focus now always shifts to modal container when opened.
+
+### Added
+* Component specific handling for pressing Escape key to close the abstract modal
+
 2.6.0 - (April 24, 2019)
 ------------------
 ### Changed