+++ packages/terra-aggregator/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+4.14.0 - (May 13, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+4.13.0 - (May 7, 2019)
+------------------
+### Changed
+* Minor dependency update
+
 4.12.0 - (April 24, 2019)
 ------------------
 ### Changed