+++ packages/terra-application-layout/CHANGELOG.md
@@ -3,8 +3,24 @@ Changelog
 
 Unreleased
 ----------
-### Fixed
-* Updated wdio screenshots to fix test failures
+
+5.1.0 - (March 15, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+5.0.0 - (May 13, 2019)
+------------------
+### Breaking Changes
+* Consume latest major version bump of terra-popup
+* Removed terra-disclosure-manager dependency
+* Added terra-disclosure-manager peer-dependency
+* Update react-router-dom dependency to latest stable major release
+
+4.12.0 - (May 7, 2019)
+------------------
+### Changed
+* Updated WDIO screenshots
 
 4.11.0 - (April 24, 2019)
 ------------------