+++ packages/terra-application-links/CHANGELOG.md
@@ -3,6 +3,19 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.should.validateElement()`
+
+6.0.0 - (May 13, 2019)
+------------------
+### Breaking Changes
+* Consume latest major version bump of terra-popup
+* Update react-router-dom peerDependency to latest stable major release
+
+5.12.0 - (May 7, 2019)
+------------------
+### Changed
+* Minor dependency update
 
 5.11.0 - (April 24, 2019)
 ------------------