+++ packages/terra-application-name/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.should.validateElement()`
+
+3.8.0 - (March 15, 2019)
+------------------
+### Changed
+* Minor dependency update
 
 3.7.0 - (April 24, 2019)
 ------------------