+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+6.0.0 - (May 13, 2019)
+------------------
+### Breaking Changes
+* Consume latest major version bump of terra-menu
+
+5.10.0 - (May 7, 2019)
+------------------
+### Changed
+* Minor dependency update
+
 5.9.0 - (April 24, 2019)
 ------------------
 ### Changed