+++ packages/terra-disclosure-manager/CHANGELOG.md
@@ -3,6 +3,14 @@ ChangeLog
 
 Unreleased
 ----------
+
+4.11.0 - (May 13, 2019)
+------------------
+### Added
+* Included documentation regarding consumption and dependency management
+
+4.10.0 - (May 7, 2019)
+------------------
 ### Added
 * Added new width 560px to Disclosure Widths.