+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -3,8 +3,16 @@ Changelog
 
 Unreleased
 ----------
+
+2.7.0 - (March 15, 2019)
+------------------
+### Added
+* Added custom font-weight property, `--terra-navigation-side-menu-item-font-weight`, for MenuItem component (defaults to 'inherit')
+
+2.6.0 - (May 7, 2019)
+------------------
 ### Fixed
-* Updated wdio screenshots to fix test failures
+* Updated WDIO screenshots
 
 2.5.0 - (April 24, 2019)
 ------------------