+++ packages/terra-popup/CHANGELOG.md
@@ -4,6 +4,18 @@ Changelog
 Unreleased
 ----------
 
+6.0.0 - (May 13, 2019)
+------------------
+### Breaking Changes
+* Removed `releaseFocus` prop
+* Removed `requestFocus` prop
+* Updated focus-trap-react dependency to ^6.0.0
+
+5.12.0 - (May 7, 2019)
+------------------
+### Changed
+* Minor dependency update
+
 5.11.0 - (April 24, 2019)
 ------------------
 ### Changed