+++ packages/terra-slide-group/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+4.0.0 - (May 13, 2019)
+------------------
+### Breaking Change
+* Updated react-transition-group dependency to latest stable major release (^4.0.0)
+
 3.3.0 - (April 24, 2019)
 ------------------
 ### Changed